organaizer/requirements.txt
aiogram<4.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
beautifulsoup4==4.12.2
pandas==2.1.2
requests==2.31.0
psycopg2-binary==2.9.9
